chore: move page and components
