+++ src/app/[locale]/(logged-in)/data-custodian/profile/[tabId]/components/Organisations/Organisations.tsx
@@ -4,18 +4,18 @@ import ContactLink from "@/components/ContactLink";
 import { Message } from "@/components/Message";
 import OverlayCenter from "@/components/OverlayCenter";
 import { useNotifications } from "@/context/Notifications";
-import PageSection from "@/modules/PageSection";
 import {
   DeleteApprovalPayloadWithEntity,
   PostApprovalPayloadWithEntity,
 } from "@/services/approvals";
 import { getOrganisations } from "@/services/organisations";
-import { CircularProgress, Typography } from "@mui/material";
+import { CircularProgress } from "@mui/material";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useTranslations } from "next-intl";
 import { useCallback } from "react";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useMutationApproval, useMutationDeleteApproval } from "../../hooks";
 import OrganisationsList from "../OrganisationsList";
+import OrganisationsLegend from "../OrganisationsLegend";
 
 const NAMESPACE_TRANSLATIONS_USERS_LIST = "UsersList";
 const NAMESPACE_TRANSLATIONS_USERS = "Users";
@@ -83,35 +83,31 @@ export default function Sections() {
 
   return (
     <>
+      <OrganisationsLegend />
       {isOrganisationsLoading && (
         <OverlayCenter variant="contained">
           <CircularProgress aria-label={tUsersList("loadingAriaLabel")} />
         </OverlayCenter>
       )}
-      <PageSection sx={{ display: "flex" }}>
-        <Typography variant="h4">{tUsersList("title")}</Typography>
-      </PageSection>
-      <PageSection sx={{ flexGrow: 1 }}>
-        {isOrganisationsError && (
-          <Message severity="error" sx={{ mb: 3 }}>
-            {tUsers.rich(`${orgainsationsError}`, {
-              contactLink: ContactLink,
-            })}
-          </Message>
-        )}
-        {!isOrganisationsLoading && organisationsData?.data.data && (
-          <OrganisationsList
-            onApprove={handleApprove}
-            onUnapprove={handleUnapprove}
-            organisations={organisationsData?.data.data}
-            mutateState={{
-              isError: isDeleteError || isUpdateError,
-              isLoading: isDeleteLoading || isUpdateLoading,
-              error: errorDelete || errorUpdate,
-            }}
-          />
-        )}
-      </PageSection>
+      {isOrganisationsError && (
+        <Message severity="error" sx={{ mb: 3 }}>
+          {tUsers.rich(`${orgainsationsError}`, {
+            contactLink: ContactLink,
+          })}
+        </Message>
+      )}
+      {!isOrganisationsLoading && organisationsData?.data.data && (
+        <OrganisationsList
+          onApprove={handleApprove}
+          onUnapprove={handleUnapprove}
+          organisations={organisationsData?.data.data}
+          mutateState={{
+            isError: isDeleteError || isUpdateError,
+            isLoading: isDeleteLoading || isUpdateLoading,
+            error: errorDelete || errorUpdate,
+          }}
+        />
+      )}
     </>
   );
 }
