+++ src/app/[locale]/(logged-in)/data-custodian/profile/[tabId]/hooks/index.ts
