--- src/app/[locale]/(logged-in)/data-custodian/users/page.tsx
@@ -1,12 +1,0 @@
-import PageContainer from "@/modules/PageContainer";
-import Sections from "./components/Sections";
-
-function Page() {
-  return (
-    <PageContainer>
-      <Sections />
-    </PageContainer>
-  );
-}
-
-export default Page;
