--- src/app/[locale]/(logged-in)/issuer/profile/[tabId]/components/Organisations/Organisations.tsx
@@ -1,5 +1,0 @@
-import OrganisationsLegend from "../OrganisationsLegend";
-
-export default function Organisations() {
-  return <OrganisationsLegend />;
-}
